fix: address import refresh review feedback
