+++ integration_test/settings_import_refresh_validation_test.dart
@@ -19,6 +19,8 @@ import 'package:monkeyssh/domain/services/settings_service.dart';
 import 'package:monkeyssh/presentation/providers/entity_list_providers.dart';
 import 'package:monkeyssh/presentation/screens/settings_screen.dart';
 
+import '../test/support/settings_import_test_helpers.dart';
+
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
@@ -31,7 +33,7 @@ void main() {
     final db = AppDatabase.forTesting(NativeDatabase.memory());
     addTearDown(db.close);
     final encryptionService = SecretEncryptionService.forTesting();
-    final transferService = _FakeSecureTransferService(
+    final transferService = FakeSecureTransferService(
       db,
       KeyRepository(db, encryptionService),
       HostRepository(db, encryptionService),
@@ -52,7 +54,7 @@ void main() {
       ),
     );
 
-    FilePicker.platform = _FakeFilePicker(
+    setFakeFilePickerResult(
       result: FilePickerResult([
         PlatformFile(
           name: 'migration.monkeysshx',
@@ -61,7 +63,6 @@ void main() {
         ),
       ]),
     );
-    addTearDown(() => FilePicker.platform = _FakeFilePicker(result: null));
 
     var hostBuilds = 0;
     var keyBuilds = 0;
@@ -71,20 +72,18 @@ void main() {
       ProviderScope(
         overrides: [
           databaseProvider.overrideWithValue(db),
-          authServiceProvider.overrideWithValue(_FakeAuthService()),
-          authStateProvider.overrideWith(_MockAuthStateNotifier.new),
+          authServiceProvider.overrideWithValue(FakeAuthService()),
+          authStateProvider.overrideWith(MockAuthStateNotifier.new),
           secureTransferServiceProvider.overrideWithValue(transferService),
-          themeModeNotifierProvider.overrideWith(_StaticThemeModeNotifier.new),
-          fontSizeNotifierProvider.overrideWith(_StaticFontSizeNotifier.new),
-          fontFamilyNotifierProvider.overrideWith(
-            _StaticFontFamilyNotifier.new,
-          ),
+          themeModeNotifierProvider.overrideWith(StaticThemeModeNotifier.new),
+          fontSizeNotifierProvider.overrideWith(StaticFontSizeNotifier.new),
+          fontFamilyNotifierProvider.overrideWith(StaticFontFamilyNotifier.new),
           cursorStyleNotifierProvider.overrideWith(
-            _StaticCursorStyleNotifier.new,
+            StaticCursorStyleNotifier.new,
           ),
-          bellSoundNotifierProvider.overrideWith(_StaticBellSoundNotifier.new),
+          bellSoundNotifierProvider.overrideWith(StaticBellSoundNotifier.new),
           terminalThemeSettingsProvider.overrideWith(
-            _StaticTerminalThemeSettingsNotifier.new,
+            StaticTerminalThemeSettingsNotifier.new,
           ),
           allHostsProvider.overrideWith((ref) {
             hostBuilds += 1;
@@ -100,7 +99,7 @@ void main() {
           }),
         ],
         child: const MaterialApp(
-          home: Stack(children: [SettingsScreen(), _EntityProviderProbe()]),
+          home: Stack(children: [SettingsScreen(), EntityProviderProbe()]),
         ),
       ),
     );
@@ -132,128 +131,4 @@ void main() {
     expect(groupBuilds, greaterThan(initialGroupBuilds));
     expect(find.text('Migration import completed'), findsOneWidget);
   });
-}
-
-class _EntityProviderProbe extends ConsumerWidget {
-  const _EntityProviderProbe();
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    ref
-      ..watch(allHostsProvider)
-      ..watch(allKeysProvider)
-      ..watch(allGroupsProvider);
-    return const SizedBox.shrink();
-  }
-}
-
-class _MockAuthStateNotifier extends AuthStateNotifier {
-  @override
-  AuthState build() => AuthState.notConfigured;
-}
-
-class _FakeAuthService extends AuthService {
-  @override
-  Future<bool> isAuthEnabled() async => false;
-
-  @override
-  Future<bool> isBiometricEnabled() async => false;
-
-  @override
-  Future<bool> isBiometricAvailable() async => false;
-}
-
-class _FakeFilePicker extends FilePicker {
-  _FakeFilePicker({required this.result});
-
-  final FilePickerResult? result;
-
-  @override
-  Future<FilePickerResult?> pickFiles({
-    String? dialogTitle,
-    String? initialDirectory,
-    FileType type = FileType.any,
-    List<String>? allowedExtensions,
-    Function(FilePickerStatus)? onFileLoading,
-    bool allowCompression = false,
-    int compressionQuality = 0,
-    bool allowMultiple = false,
-    bool withData = false,
-    bool withReadStream = false,
-    bool lockParentWindow = false,
-    bool readSequential = false,
-  }) async => result;
-}
-
-class _FakeSecureTransferService extends SecureTransferService {
-  _FakeSecureTransferService(
-    super.db,
-    super.keyRepository,
-    super.hostRepository, {
-    required this.payload,
-  });
-
-  final TransferPayload payload;
-  int importCallCount = 0;
-
-  @override
-  Future<TransferPayload> decryptPayload({
-    required String encodedPayload,
-    required String transferPassphrase,
-  }) async => payload;
-
-  @override
-  MigrationPreview previewMigrationPayload(TransferPayload payload) =>
-      MigrationPreview(
-        settingsCount: 0,
-        hostCount: (payload.data['hosts'] as List).length,
-        keyCount: (payload.data['keys'] as List).length,
-        groupCount: (payload.data['groups'] as List).length,
-        snippetCount: (payload.data['snippets'] as List).length,
-        snippetFolderCount: (payload.data['snippetFolders'] as List).length,
-        portForwardCount: (payload.data['portForwards'] as List).length,
-        knownHostCount: (payload.data['knownHosts'] as List).length,
-      );
-
-  @override
-  Future<void> importFullMigrationPayload({
-    required TransferPayload payload,
-    required MigrationImportMode mode,
-  }) async {
-    importCallCount += 1;
-  }
-}
-
-class _StaticThemeModeNotifier extends ThemeModeNotifier {
-  @override
-  ThemeMode build() => ThemeMode.system;
-}
-
-class _StaticFontSizeNotifier extends FontSizeNotifier {
-  @override
-  double build() => 14;
-}
-
-class _StaticFontFamilyNotifier extends FontFamilyNotifier {
-  @override
-  String build() => 'System Monospace';
-}
-
-class _StaticCursorStyleNotifier extends CursorStyleNotifier {
-  @override
-  String build() => 'block';
-}
-
-class _StaticBellSoundNotifier extends BellSoundNotifier {
-  @override
-  bool build() => true;
-}
-
-class _StaticTerminalThemeSettingsNotifier
-    extends TerminalThemeSettingsNotifier {
-  @override
-  TerminalThemeSettings build() => const TerminalThemeSettings(
-    lightThemeId: 'github-light',
-    darkThemeId: 'dracula',
-  );
 }
