+++ lib/presentation/providers/entity_list_providers.dart
@@ -1,4 +1,5 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:flutter_riverpod/misc.dart' show ProviderBase;
 
 import '../../data/database/database.dart';
 import '../../data/repositories/group_repository.dart';
@@ -23,10 +24,13 @@ final allGroupsProvider = StreamProvider<List<Group>>((ref) {
   return repo.watchAll();
 });
 
+/// Signature for invalidating shared providers from any Riverpod context.
+typedef ProviderInvalidator =
+    void Function(ProviderBase<Object?> provider, {bool asReload});
+
 /// Refreshes shared entity list providers after migration imports replace data.
-void invalidateImportedEntityProviders(WidgetRef ref) {
-  ref
-    ..invalidate(allHostsProvider)
-    ..invalidate(allKeysProvider)
-    ..invalidate(allGroupsProvider);
+void invalidateImportedEntityProviders(ProviderInvalidator invalidate) {
+  invalidate(allHostsProvider);
+  invalidate(allKeysProvider);
+  invalidate(allGroupsProvider);
 }
